Fixed admin to prevent orders without items, recalculate shipping price and send updates if modifications apply

+++ app/admin/admin_prana_orders.rb
@@ -6,7 +6,7 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
   filter :order_number, label: "Número de orden"
   filter :users, collection: -> { User.all.map { |user| [user.external_id, user.id] }.sort }
 
-  permit_params :description, :order_number, :user_ids, :created_at, item_ids: [], items_attributes: [:id]
+  permit_params :description, :order_number, :user_ids, :created_at, :shipping_address_id, :shipping_price, item_ids: [], items_attributes: [:id, :item, :_destroy]
 
   controller do
 
@@ -15,31 +15,82 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
     end
 
     def create
-      params[:order][:item_ids] = []
-      items = params[:order][:items_attributes]
-      items.each do |ix, item|
-        params[:order][:item_ids] << item[:id]
+      if params[:order][:items_attributes]
+        params[:order][:item_ids] = []
+        items = params[:order][:items_attributes]
+        items.each do |ix, item|
+          params[:order][:item_ids] << item[:id]
+        end
+        params[:order].delete("items_attributes")
       end
-      params[:order].delete("items_attributes")
+      
+      if params[:order][:created_at].to_datetime.offset == 0
+        params[:order][:created_at] = params[:order][:created_at].in_time_zone.to_s 
+      end
+
       super
     end
 
     def update
-      params[:order][:item_ids] = []
-      items = params[:order][:items_attributes]
-      items.each do |ix, item|
-        if item[:_destroy] == "1"
-          next
+      if params[:order][:items_attributes] 
+        params[:order][:item_ids] = []
+        items = params[:order][:items_attributes]
+        destroy_count = item_count = 0
+        items.each do |ix, item|
+          item_count += 1
+          if item[:_destroy] == "1"
+            destroy_count += 1
+            next
+          end
+          params[:order][:item_ids] << item[:id]
         end
-        params[:order][:item_ids] << item[:id]
-      end
-      params[:order].delete("items_attributes")
+        if destroy_count == item_count
+          raise 'No pueden haber ordenes sin items'
+        end
+        params[:order].delete("items_attributes")
 
-      order = Order.find(params[:id])
-      order.items.destroy_all
-      params[:order][:item_ids].each do |item_id|
-        order.items << Item.find(item_id)    
+        order = Order.find(params[:id])
+        original_item_ids = [] 
+        order.items.each do |item| 
+          original_item_ids << item.id.to_s
+        end
+
+        order.items.destroy_all
+        params[:order][:item_ids].each do |item_id|
+          order.items << Item.find(item_id)    
+        end
       end
+
+      email_update, update_volume = false
+      if order.created_at.to_s != params[:order][:created_at] 
+        if params[:order][:created_at].to_datetime.offset == 0
+          params[:order][:created_at] = params[:order][:created_at].in_time_zone.to_s 
+        end
+        
+        email_update = true
+        update_volume = true
+        order.update_column(:created_at, params[:order][:created_at]) 
+      end
+      if params[:order][:item_ids].sort != original_item_ids.sort
+        email_update = true
+        update_volume = true
+      end
+      if params[:order][:shipping_address_id] != order.shipping_address_id.to_s
+        email_update = true
+        order.update_column(:shipping_address_id, params[:order][:shipping_address_id])
+      end
+      if params[:order][:shipping_price].to_f != order.shipping_price
+        email_update = true
+        order.update_column(:shipping_price, params[:order][:shipping_price])
+        order.update_column(:total_price,  order.calculate_total_price)
+      end
+      if email_update
+        SendEmailJob.perform_later("order", order.users.first, order)
+      end
+      if update_volume
+        order.update_volume_for_users
+      end
+
       params[:order].delete("item_ids")
       super
     end
@@ -57,6 +108,9 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
     column "Nombre" do |order|
       "#{User.find(order.user_ids.first).first_name} #{User.find(order.user_ids.first).last_name}"
     end
+    column "Dirección" do |order|
+      order.shipping_address.to_s
+    end
     column "Item" do |order|
       items = ""
       order.item_ids.each do |item_id|
@@ -66,12 +120,18 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
       items.html_safe
     end
     column "Puntos" do |order|
-      volume = 0
-      order.item_ids.each do |item_id|
-        item = Item.find(item_id)
-        volume += item.volume
+      order.total_item_volume
+    end
+    column "Precio de productos" do |order|
+      order.total_item_price
+    end
+    column "Precio de envío", :shipping_price 
+    column "Precio total" do |order|
+      if order.total_price
+        order.total_price
+      else
+        order.calculate_total_price
       end
-      volume
     end
     
     actions defaults: true
@@ -106,33 +166,21 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
           .map{|user| ["#{user.external_id} - #{user.first_name} #{user.last_name}", user.id]}, 
           include_blank: false
       end
-      
-      #f.input :item_ids, label: "Item", as: :select, collection: Item.all.map {|item| ["#{item.company}-#{item.name}", item.id]}.sort,
-      #  include_blank: false
 
-      #item_collection = Item.all.map {|item| ["#{item.company}-#{item.name}", item.id]}.sort
-      #item_collection += Item.all.map {|item| ["#{item.company}-#{item.name}", item.id]}.sort
-      #item_collection += Item.all.map {|item| ["#{item.company}-#{item.name}", item.id]}.sort
-      #item_collection += Item.all.map {|item| ["#{item.company}-#{item.name}", item.id]}.sort
-      #f.input :items, as: :select, collection: Item.all.map {|item| ["#{item.company}-#{item.name}", item.id]}.sort, include_blank: false
-      #f.input :item_ids, label: "Productos", as: :check_boxes, collection: item_collection, 
-      #  include_blank: false
-      
-      #f.input :item, as: :select, collection: Item.all.map {|item| "#{item.company}-#{item.name}"}.sort, include_blank: false
-
-      #f.inputs "Usuario" do
-      #  f.has_many :users, new_record: false do |a|
-      #    a.input :external_id, label: "ID Omein", input_html: { disabled: true, style: "background-color: #d3d3d3;" }
-      #  end
-      #end
-      omein_item_collection = Item.where(company: PranaCompPlan::COMPANY_PRANA).map {|item| ["#{item.name}", item.id]}.sort
-      f.inputs "Productos" do
-        f.has_many :items, new_record: true, allow_destroy: true do |a|
-         a.input :id, as: :select, collection: omein_item_collection, include_blank: false
+      if not f.object.users.empty?
+        f.inputs "Dirección de envío" do
+          f.input :shipping_address, label: "Dirección", as: :select, 
+            collection: f.object.users.first.shipping_addresses.map{|sa| [sa.to_s, sa.id]}
         end
       end
-      #f.input :user_ids, as: :select, collection: User.all.map {|user| user.external_id}, include_blank: false
-      #f.input :item_ids, as: :select, collection: Item.all.map {|item| "#{item.company}-#{item.name}"}.sort, include_blank: false
+      
+      prana_item_collection = Item.where(company: PranaCompPlan::COMPANY_PRANA, active: true).map {|item| ["#{item.name}", item.id]}.sort
+      f.inputs "Productos" do
+        f.has_many :items, new_record: true, allow_destroy: true do |a|
+         a.input :id, as: :select, collection: prana_item_collection, include_blank: false
+        end
+      end
+      f.input :shipping_price, label: "Precio de envío" 
     end
     f.actions   
 
@@ -157,6 +205,9 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
     column "Nombre" do |order|
       "#{User.find(order.user_ids.first).first_name} #{User.find(order.user_ids.first).last_name}"
     end
+    column "Direccion" do |order|
+      order.shipping_address.to_s      
+    end
     column "Item" do |order|
       items = ""
       order.item_ids.each do |item_id|
@@ -166,12 +217,20 @@ ActiveAdmin.register Order, as: "Prana Ordenes" do
       items.html_safe
     end
     column "Puntos" do |order|
-      volume = 0
-      order.item_ids.each do |item_id|
-        item = Item.find(item_id)
-        volume += item.volume
+      order.total_item_volume
+    end
+    column "Precio de productos" do |order|
+      order.total_item_price
+    end
+    column "Precio de envio" do |order|
+      order.shipping_price
+    end
+    column "Precio total" do |order|
+      if order.total_price
+        order.total_price
+      else
+        order.calculate_total_price
       end
-      volume
     end
   end
